python-service/requirements.txt
flask>=3.1.0
openai-whisper>=20250625
torch>=2.7.0
torchaudio>=2.7.0
werkzeug>=3.1.0
numpy>=1.24,<2.3
nltk>=3.9.1
scikit-learn>=1.7.0
spacy>=3.8.0
numba>=0.61.0
moviepy>=1.0.3
